requirements.txt
Django==1.11
django-cors-headers
mysqlclient
djangorestframework
